+++ src/MeshManager.h
@@ -2,6 +2,7 @@
 #define MESHMANAGER_H
 
 #include <string>
+#include <sys/types.h>
 
 namespace SimpleGL {
   class MeshManagerPrivate;
@@ -17,7 +18,7 @@ namespace SimpleGL {
     Mesh *createQuad(float width = 1.0f, float height = 1.0f);
     Mesh *createPlane(float width = 1.0f, float height = 1.0f, float uTile = 1.0f, float vTile = 1.0f);
     Mesh *createCube(float width = 1.0f, float height = 1.0f, float depth = 1.0f);
-    Mesh *createSphere(float radius = 1.0f, float slices = 16.0f, float stacks = 16.0f);
+    Mesh *createSphere(float radius = 1.0f, uint slices = 16, uint stacks = 16);
 
     Mesh *loadMesh(const std::string &path);
